requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
qrcode==7.4.2
Pillow==10.2.0
gunicorn==21.2.0
pytest==8.2.1
